Migrate tests to regression, add deprecation
