+++ tests/test_filterset.py
@@ -2,8 +2,6 @@
 from __future__ import absolute_import
 from __future__ import unicode_literals
 
-import datetime
-
 import django
 from django.test import TestCase
 
@@ -22,7 +20,6 @@ from .testapp.filters import (
     NoteFilterWithRelatedAll,
     NoteFilterWithRelatedAllDifferentFilterName,
     PostFilter,
-    CoverFilterWithRelatedMethodFilter,
     CoverFilterWithRelated,
     # PageFilterWithRelated,
     TagFilter,
@@ -554,44 +551,6 @@ class FilterSubsetTests(TestCase):
         # non-filter params should be ignored
         filterset_class = NoteFilterWithRelated.get_subset(['foobar'])
         self.assertEqual(len(filterset_class.base_filters), 0)
-
-
-class MethodFilterTests(TestCase):
-
-    @classmethod
-    def setUpTestData(cls):
-        user = User.objects.create(username="user1", email="user1@example.org")
-
-        note1 = Note.objects.create(title="Test 1", content="Test content 1", author=user)
-        note2 = Note.objects.create(title="Test 2", content="Test content 2", author=user)
-
-        post1 = Post.objects.create(note=note1, content="Test content in post 1")
-        post2 = Post.objects.create(note=note2, content="Test content in post 2", date_published=datetime.date.today())
-
-        Cover.objects.create(post=post1, comment="Cover 1")
-        Cover.objects.create(post=post2, comment="Cover 2")
-
-    def test_method_filter(self):
-        GET = {
-            'is_published': 'true'
-        }
-        filterset = PostFilter(GET, queryset=Post.objects.all())
-        results = list(filterset.qs)
-        self.assertEqual(len(results), 1)
-        self.assertEqual(results[0].content, "Test content in post 2")
-
-    def test_related_method_filter(self):
-        """
-        Missing MethodFilter filter methods are silently ignored, returning
-        the unfiltered queryset.
-        """
-        GET = {
-            'post__is_published': 'true'
-        }
-        filterset = CoverFilterWithRelatedMethodFilter(GET, queryset=Cover.objects.all())
-        results = list(filterset.qs)
-        self.assertEqual(len(results), 1)
-        self.assertEqual(results[0].comment, "Cover 2")
 
 
 class FilterOverrideTests(TestCase):
